+++ packages/terra-aggregator/src/terra-dev-site/test/aggregator/AggregatorWithDisclosure.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import SlidePanelManager from 'terra-slide-panel-manager';
 import { withDisclosureManager } from 'terra-disclosure-manager';
 import classNames from 'classnames/bind';
-import styles from './AggregatorWithDisclosure.test.scss';
+import styles from './AggregatorWithDisclosure.test.module.scss';
 
 import Aggregator from '../../../Aggregator';
 import AggregatorItem from '../common/AggregatorItem';