+++ packages/terra-dialog-modal/src/terra-dev-site/doc/example/DialogModalWithCustomHeaderAndCustomFooter.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Button from 'terra-button';
 import classNames from 'classnames/bind';
 import DialogModal from '../../../DialogModal';
-import styles from './DialogModalWithCustomHeaderAndCustomFooter.scss';
+import styles from './DialogModalWithCustomHeaderAndCustomFooter.module.scss';
 
 const cx = classNames.bind(styles);