+++ packages/terra-slide-group/src/terra-dev-site/doc/example/SlideGroupDemo.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import SlideGroup from 'terra-slide-group/lib/SlideGroup';
-import styles from './SlideGroupDemo.scss';
+import styles from './SlideGroupDemo.module.scss';
 
 const cx = classNames.bind(styles);