+++ packages/terra-theme-provider/src/terra-dev-site/doc/example/MockThemeComponent.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
-import styles from 'terra-theme-provider/lib/terra-dev-site/doc/example/MockThemeComponent.scss';
+import styles from 'terra-theme-provider/lib/terra-dev-site/doc/example/MockThemeComponent.module.scss';
 import 'terra-theme-provider/lib/terra-dev-site/theme/scoped-theme.scss';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */