+++ packages/terra-infinite-list/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Removed extra inline styles.
+* Renamed test and example SCSS files to end with `.module.scss` extension
 
 3.7.0 - (June 19, 2019)
 ------------------