+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Update the modal manager test example to not test accessibility for elements behind an overlay
+* Renamed test and example SCSS files to end with `.module.scss` extension
 
 3.3.0 - (June 19, 2019)
 ------------------